@@ -1,13 +1,16 @@
 package la321.katy.com.fatcat.irelevant;
 
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
+
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import la321.katy.com.fatcat.views.fragments.NoFragment;
 import la321.katy.com.fatcat.views.fragments.YesFragment;
 
-public class ResultPagerAdapter extends android.support.v4.app.FragmentPagerAdapter {
+public class ResultPagerAdapter extends FragmentPagerAdapter {
     public ResultPagerAdapter(FragmentManager fm) {
         super(fm);
     }
