@@ -2,25 +2,26 @@ package la321.katy.com.fatcat.views;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
+import androidx.appcompat.app.AppCompatActivity;
+
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
 import la321.katy.com.fatcat.App;
 import la321.katy.com.fatcat.R;
 
 public class AboutActivity extends AppCompatActivity {
 
-    @BindView(R.id.fab_move_to_pet_prop)
+
     FloatingActionButton moveToPetProp;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_about);
-        ButterKnife.bind(this);
+        moveToPetProp = findViewById(R.id.fab_move_to_pet_prop);
         moveToPetProp.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
