@@ -1,7 +1,9 @@
 package la321.katy.com.fatcat.views;
 
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import la321.katy.com.fatcat.R;
 
