@@ -2,10 +2,11 @@ package la321.katy.com.fatcat.views.fragments;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 import la321.katy.com.fatcat.R;
 
