@@ -2,11 +2,12 @@ package la321.katy.com.fatcat.views.fragments;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+
+import androidx.fragment.app.Fragment;
 
 import la321.katy.com.fatcat.R;
 import la321.katy.com.fatcat.views.LoginActivity;
